+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Added
 - Task error_summary and error_details are now initialized appropriately
   with data from Pulp
+- Client now logs Pulp load every few minutes when awaiting tasks
 
 ## [0.1.1] - 2019-06-13